fix: action confirm modal
