+++ app/frontend/src/components/modals/confirm-action/index.tsx
@@ -8,63 +8,58 @@ import {
 } from '@nextui-org/react';
 import { closeModals } from '../../../actions/modal';
 import useModalsInfo from '../../../hooks/use-modals-info';
-import { DesktopAPI } from '../../../desktop';
+import { useMemo } from 'react';
 
-type TVersionMismatchModalProps = {
-  clientVersion: string;
-  serverVersion: string;
+type TConfirmActionModalProps = {
+  onCancel?: () => void;
+  onConfirm?: () => void;
+  title?: string;
+  message?: string | React.ReactNode;
+  confirmLabel?: string;
+  cancelLabel?: string;
+  variant?: 'danger' | 'info';
 };
 
-const VersionMismatchModal = ({
-  clientVersion,
-  serverVersion
-}: TVersionMismatchModalProps) => {
+const ConfirmActionModal = ({
+  onCancel,
+  onConfirm,
+  title,
+  message,
+  confirmLabel,
+  cancelLabel,
+  variant
+}: TConfirmActionModalProps) => {
   const { isModalOpen } = useModalsInfo();
+  const buttonColor = useMemo(
+    () => (variant === 'danger' ? 'danger' : 'primary'),
+    [variant]
+  );
 
   return (
     <Modal
       backdrop="blur"
       isOpen={isModalOpen}
       onClose={() => {
+        onCancel?.();
         closeModals();
       }}
       scrollBehavior="inside"
     >
       <ModalContent>
         <ModalHeader className="flex gap-1 items-center">
-          <p>Version Mismatch</p>
+          <p>{title ?? 'Please confirm your action.'}</p>
         </ModalHeader>
-        <ModalBody>
-          <p>
-            The server version is different from the client version. This may
-            cause issues. Please make sure the server and this app are up to
-            date.
-          </p>
-          <div className="flex flex-col">
-            <div>
-              <span className="text-gray-500">Server version:</span>
-              <span className="ml-1">{serverVersion}</span>
-            </div>
-            <div>
-              <span className="text-gray-500">Client version:</span>
-              <span className="ml-1">{clientVersion}</span>
-            </div>
-          </div>
-
-          <span
-            className="text-blue-500 hover:underline cursor-pointer"
-            onClick={() =>
-              DesktopAPI.openUrl(
-                'https://github.com/diogomartino/palworld-ds-gui/releases/latest'
-              )
-            }
-          >
-            Download latest versions here
-          </span>
-        </ModalBody>
+        <ModalBody>{message ?? 'Are you sure?'}</ModalBody>
         <ModalFooter className="justify-center">
-          <Button variant="shadow" color="primary" onClick={closeModals}>
-            I understand
+          <Button variant="ghost" onClick={() => onCancel?.()}>
+            {cancelLabel ?? 'Cancel'}
+          </Button>
+          <Button
+            variant="solid"
+            onClick={() => onConfirm?.()}
+            color={buttonColor}
+          >
+            {confirmLabel ?? 'Confirm'}
           </Button>
         </ModalFooter>
       </ModalContent>
@@ -72,4 +67,4 @@ const VersionMismatchModal = ({
   );
 };
 
-export default VersionMismatchModal;
+export default ConfirmActionModal;
